requirements.txt
Pillow==10.2.0
tqdm==4.66.1
colorama==0.4.6